feat: Start Render component

+++ client/include/GUI/Font.hpp
@@ -8,7 +8,7 @@
 #pragma once
 
 #include <map>
-#include <ft2build.h>
+#include "ft2build.h"
 
 #include "glm.hpp"
 
@@ -69,7 +69,7 @@ namespace LE {
                     unsigned int TextureID; // ID handle of the glyph texture
                     glm::ivec2   Size;     // Size of glyph
                     glm::ivec2   Bearing;   // Offset from baseline to left/top of glyph
-                    unsigned int Advance;   // Offset to advance to next glyph
+                    long int Advance;   // Offset to advance to next glyph
                 };
 
                 /**
